requirements.txt
openai
langchain
python-dotenv
chromadb
tiktoken
psycopg2
flask
flask-cors